// src/types/auth.ts
export type ClaimYourAccountBody = {
  work_email: string;
  first_name: string;
  last_name: string;
  password: string;
  loan_urgency: 'yesterday' | 'this-month' | '3-months' | 'this-year';
};